improve action system to see if its connected to reducer
