+++ src/actions/index.js
@@ -1,45 +1,44 @@
 /* eslint-disable */
+const CREATE_TOKEN = 'CREATE_TOKEN';
+const REMOVE_TOKEN = 'REMOVE_TOKEN';
+const CREATE_CONSULTANCIES_LIST = 'CREATE_CONSULTANCIES_LIST';
+const CREATE_APPOINTMENTS_LIST = 'CREATE_APPOINTMENTS_LIST';
+const SET_APPOINTMENT = 'SET_APPOINTMENT';
+const REMOVE_APPOINTMENT = 'REMOVE_APPOINTMENT';
+
 // import axios from 'axios';
-import BASE from '../services/appointmentapi';
-import * as types from './types';
 
-const createToken = options => {
-  console.log('1');
-  return async dispatch => {
-    console.log('2');
-    try {
-      const response = await fetch(`${BASE}/signup`, options);
-      const data = await response.json();
-      const auth = {
-        authToken: data.auth_token,
-      };
-      console.log(response);
-      console.log(auth);
-      dispatch(createTokenSuccess(auth))
-    } catch (error) {
-      console.log(error);
-    }
-  }
-};
-const createTokenSuccess = (data) => {
-  return {
-    type: types.CREATE_TOKEN,
-    payload: data,
-  }
-}
-// const createToken = async options => {
-//   const response = await fetch(`${BASE}/signup`, options);
-//   const data = await response.json();
-//   const auth = {
-//     authToken: data.auth_token,
-//   };
 
-//   console.log(types);
-//   return {
-//     type: types.CREATE_TOKEN,
-//     payload: auth,
+// const createToken = options => {
+//   console.log('1');
+//   return async dispatch => {
+//     console.log('2');
+//     try {
+//       const response = await fetch(`${BASE}/signup`, options);
+//       const data = await response.json();
+//       const auth = {
+//         authToken: data.auth_token,
+//       };
+//       console.log(response);
+//       console.log(auth);
+//       dispatch(createTokenSuccess(auth))
+//     } catch (error) {
+//       console.log(error);
+//     }
 //   }
 // };
+// const createTokenSuccess = (data) => {
+//   return {
+//     type: types.CREATE_TOKEN,
+//     payload: data,
+//   }
+// }
+const createToken = auth => {
+  return {
+    type: CREATE_TOKEN,
+    payload: auth,
+  }
+};
 
 const removeToken = () => ({
   type: REMOVE_TOKEN,
@@ -66,6 +65,7 @@ const removeAppointment = id => ({
   payload: id,
 });
 export {
+  CREATE_TOKEN,
   createToken,
   createConsultanciesList,
   createAppointmentsList,
