+++ src/containers/SignUp.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
@@ -6,11 +5,10 @@ import SignUpForm from '../components/signUpForm';
 import signUpRequest from '../requests/signUpRequest';
 import { createToken, testAction } from '../actions/index';
 
-export const SignUp = (props) => {
-  console.log(props)
+const SignUp = ({ testAction, createToken }) => {
   const handleSubmit = data => {
-    // testActionASDASD();
-    // signUpRequest(createToken, data);
+    testAction();
+    signUpRequest(createToken, data);
   };
   return (
     <div>
@@ -24,8 +22,9 @@ SignUp.propTypes = {
   createToken: PropTypes.func,
   testAction: PropTypes.func,
 };
+
 SignUp.defaultProps = {
-  // createToken,
-  // testAction,
+  createToken,
+  testAction,
 };
 export default connect(null, { testAction })(SignUp);
