+++ src/store/reducers/authenticationReducer.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-import * as types from '../../actions/types';
+import { CREATE_TOKEN } from '../../actions/index';
 
 const defaultState = {
   name: '',
@@ -8,12 +8,11 @@ const defaultState = {
   authToken: '',
 };
 const authenticationReducer = (state = defaultState, action) => {
-  console.log(action);
   // debugger
   switch (action.type) {
-    case types.CREATE_TOKEN:
+    case CREATE_TOKEN:
       console.log(action, 'ddsadw');
-      return { ...state, authToken: action.payload };
+      return action.payload;
     case 'TEST':
       console.log('TEST REDUCER WORLKED');
       return { ...state, test: 'test worked' };
